pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-environ>=0.12.0",
    "django-phonenumber-field[phonenumberslite]>=8.1.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg>=3.2.10",
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]
